feat: Update version numbers and add product version
This commit updates the product version to 0.1.0 and updates the AsvMavlink version to 3.2.5-alpha-16 in Directory.Build.props. Alongside this, assembly versions and file versions are further added to various project files such as Asv.Drones.Gui.Uav, Asv.Drones.Gui, etc., for better management and tracking of each version. These changes are made to keep the project files up-to-date and well tracked with the newer version updates, ensuring ease in managing changes in disparate project files.