fix: update .net version in github workflows
Updates the .Net version used in the setup for MacOS, Android, Ubuntu, and Windows release workflows. The version has been upgraded from 6.x.x to 7.x.x.

This change is necessary to align with the latest features of .Net framework and improve the overall functionality and efficiency of the application.

Asana: https://app.asana.com/0/1203851531040615/1204953971605260/f