fix(core): update author and app URLs in CorePlugin.cs
Adjusted the author and application URLs in CorePlugin.cs to match changes in website domains. The author URL is changed from the Github profile to a personal website, and the app URL is now directing to the documentation site instead of the Github repository. This is to provide more accurate and helpful information to the users.

Asana: https://app.asana.com/0/1203851531040615/1204953971605265/f