fix: add Android CI workflow and badge to README
Added a line to the Github Android release workflow file to install the android workload. This change ensures that the Android dotnet tools are installed in the CI server, enabling the Android release workflow to run successfully. Also, updated the README to include a status badge for the Android CI workflow, providing clear visibility of the workflow's current status.

Asana: https://app.asana.com/0/1203851531040615/1204953971605260/f