bug(core): remove file dialogs dependency and remove dynamic app store
On mobile platforms, there is no "File" and "Folder" dialog boxes. We can remove this functionality. In particular, we can't change theAppStore at runtime (because we don't have file open dialogs). If this is the case, it will be easier to develop a program without dynamic storage.

Asana: https://app.asana.com/0/1203851531040615/1204953971605303/f