Add VorSdrRtt module's ViewModel and View
This commit added the ViewModel and View of VorSdrRtt module in the Asv.drones Gui Service. Updates include adding new properties to the ViewModel for power, frequency offset, bearing, modulation amplitude, and frequency. ViewModel properties are then used in the corresponding View to display data, achieving unidirectional data flow. This change is necessary to improve the functionality of the VorSdrRtt module and ensure data is accurately represented in the UI

Asana: https://app.asana.com/0/1203851531040615/1204962408289039/f