feat(uav): add reload controller feature in GUI
This commit adds a new feature in the UAV GUI shell which allows for the reloading of a controller. This is facilitated by integration of a new 'Reload' button, which calls a command in the ViewModel. Moreover, the change also comprises of modifications in the resource files to incorporate the necessary localized strings for the 'Reload' button content. This enhancement ensures that the user can easily reset the controller, drastically improving user experience and system performance in real-time scenarios.

Asana: https://app.asana.com/0/1203851531040615/1204860778906959/f