feat(uav): added UavAction class and actions list to FlightUavViewModel
A new class called UavAction was introduced to encapsulate UAV-related actions with their commands and titles. These actions are now kept in a list in the FlightUavViewModel. The UI was updated accordingly to list these actions in a drop-down menu for the user. This enhances the user interaction by grouping all relevant actions together and provides a convenient interface for adding and handling future actions.

Asana: https://app.asana.com/0/1203851531040615/1205007794960752/f