fix: update vehicle current mode dynamically
Modified the Vehicle.CurrentMode initialization way in both FlightUavViewModel and SelectModeAnchorActionViewModel files. Instead of initializing it with the vehicle's current mode during instantiation time, it now subscribes to the Vehicle.CurrentMode changes. It ensures that the CurrentMode on the UI will be updated whenever the Vehicle's CurrentMode gets updated in the backend.

This change was necessary to keep the UI view synchronized with the backend vehicle current mode, especially when it changes after the object creation. Thus, users will always see the current and actual mode of the vehicle in the UI.

Asana: https://app.asana.com/0/1203851531040615/1204961236602308/f