fix: change default ruler visibility upon map loading
The default state of the ruler when the map is loaded has been changed from visible to hidden. This is to improve the initial user experience by reducing clutter and promoting a cleaner UI. Further interactions with the ruler can be made by the user as needed.

Asana: https://app.asana.com/0/1203851531040615/1205041729023507/f