fix(map): update polygon points in MapView.axaml
Updated the points for the polygon elements in MapView.axaml to be dynamically retrieved from the LeftCornerPoints and RightCornerPoints properties in the MapView.axaml.cs. This enables the polygonal shape to be changed dynamically through the respective properties. Also added a dialog mode in design mode for better UI representation.

Asana: https://app.asana.com/0/1203851531040615/1204967469316981/f