fix(uav): update Map follow state for UAV widgets
The map's UAV widget follow state was not being properly updated when the map item to follow was set to null. This change ensures that all of the UAV widgets are set to 'not followed' when the map item to follow is made null. This will prevent inconsistencies in the follow-status between the map and the UAV widgets.

Asana: https://app.asana.com/0/1203851531040615/1205044822709643/f