feat (core): add AnchorMover functionality in Map Actions
This commit introduces functionality to move anchors in the Map view. The code changes include the addition of AnchorMoverActionViewModel, its view, and updates to the .DotSettings file to skip namespace validation for the new files added. This is to enhance the map editing experience by enabling users to easily adjust the positions of anchors.

Asana: https://app.asana.com/0/1203851531040615/1204998445929976/f