feat (Map): Add anchor edit mode for map functionality
The anchor edit mode has been introduced for the MapView component. This change specifically adds a DirectProperty in the MapView class named 'IsInAnchorEditMode'. With the inclusion of the anchor edit mode, additional checks for `_map.IsInAnchorEditMode` were added in MapViewItem to ensure its correct usage. Also, in MapPageView, 'IsInAnchorEditMode' is bound to 'IsInEditMode' to reflect proper state changes across the UI. A static initializer is added in MapAnchorView to handle change in `IsInEditModeProperty` and reflect it in the UI using PseudoClasses. `IsInEditMode` is also added in `IMapAnchorViewModel` for consistency. This change will allow users to edit Anchors within the MapView component.

Asana: https://app.asana.com/0/1203851531040615/1204998445929976/f