feat(core): add angle units and version bump
Updated AsvCommonVersion to 1.9.0 in Directory.Build.props.
Renamed Angle.cs to Degrees.cs and refactored according class and its dependencies for better understanding and management of angle units in the UI.
Resource strings for angle measurement units added in Russian and English to provide multilingual support for the software. The proprietary parameter was removed in the Debug configuration in Asv.Drones.Gui.csproj, and a superfluous directory was removed from Asv.Drones.Gui.Core.csproj.
A localization service was added for DegreeUnits, and the ability for users to change units of measure for angle in the settings page was implemented.

Asana: https://app.asana.com/0/1203851531040615/1205080848469722/f