fix(core): add theme loading on ShellView load
An `OnLoaded` event handler has been added to `ShellView.axaml.cs`. This allows `ShellViewModel` to load the theme from the config once the view is loaded. The `IThemeService` and `IConfiguration` dependencies were also added to the `ShellViewModel` constructor. This ensures that the preferred theme is applied when the ShellView is being rendered.

Asana: https://app.asana.com/0/1203851531040615/1204953970281515/f