feat (sdr): update frequency validation in FlightSdrViewModel
Changed the frequency validation logic in `FlightSdrViewModel.cs` to directly validate the frequency input as a number greater than zero, instead of using _freqInMHzMeasureUnit.IsValid(_). This was done to simplify and improve the speed of validation, removing the need to convert and validate the input through the measure unit methods.

Asana: https://app.asana.com/0/1203851531040615/1205134815539121/f