feat(core): Remove CoordinateCalculator and introduce Templater
CoordinateCalculatorHelper.cs has been removed as it was no longer necessary for the project. Instead, a new feature, Templater, has been introduced. This feature provides an interface for creating templates with integrated tools for image and string tagging and the ability to dynamically adjust tables. It is purpose-built to replace any specific tags with the provided text in headers, footers, and the document body and is useful for functions such as template placeholders. The user interface for Templater is also implemented in TemplaterView.axaml, and the functions are in TemplaterViewModel.cs.

Asana: https://app.asana.com/0/1203851531040615/1205133534920134/f