fix(sdr): set fixed width for RecordStartView dialog
In this commit, I've enforced the width of the RecordStartView dialog to be 450 units. The Minimum and Maximum widths were also set to 450 units to ensure the dialog's size remains consistent across different user interactions. In addition, the button function in RecordStartViewModel was updated to only trigger if the information provided is valid. This is essential for preventing the application from accepting incorrect inputs, enhancing overall user experience and application's stability.

Asana: https://app.asana.com/0/1203851531040615/1205217582291092/f