fix(sdr): add FlightSdrViewModelConfig to persist frequency settings
The FlightSdrViewModel has been updated to include a FlightSdrViewModelConfig.
This change, which also includes the parsing of numeric values to be culture-independent, allows the frequency settings to be persisted and reused across sessions, which enhances the usability of the software. This change checks the validity of the frequency inputted and sets the configuration with the new valid frequency when the Mode is updated.

Asana: https://app.asana.com/0/1203851531040615/1205265909087211/f