fix: remove unused FlightSdrViewModelConfig class
The FlightSdrViewModelConfig class was deleted from the FlightSdrViewModel file. This class was previously used to store settings, but it was not used anywhere, as the FrequencyInMhz property was read once at initialization and then was lost. This resulted in unnecessary code complexity and possible confusion. By removing this class, the code becomes easier to understand and maintain.

Asana: https://app.asana.com/0/1203851531040615/1205265909087211/f