fix(sdr): refactor UI elements in the SDR flight view
The previous implementation of the UI for selecting channels in the SDR flight view was cluttered and ambiguous. This change replaces the ToggleButton with a DropDownButton, which is a more intuitive UI component. Now, when the user clicks the DropDownButton (represented by the Table icon), a list of available channels will appear, from which the user can select and the channel's respective frequency will be displayed in the TextBox. This implementation is more efficient and user-friendly. The 'IsChannelInput' reactive property was removed as it's no longer needed with the new UI design.

Asana: https://app.asana.com/0/1203851531040615/1205307075789323/f