fix(sdr): improve handling of failed tag setup in Flight Sdr
This change enhances the tag setup process in Flight Sdr ViewModel. Previously, for the types LongTag, ULongTag, DoubleTag, and StringTag, if a tag setup failed during any of the five attempts, it just broke from the loop. Now, if the setup is unsuccessful, it doesn't just break from the loop but returns to prevent further execution and logs an error message indicating the specific tag name and setup result. This helps in easier identification and debugging of the specific tag setup failures.

Asana: https://app.asana.com/0/1203851531040615/1204564719763918/f