fix(sdr): update language resources and UI for clarity
Updated some language resource files (both English and Russian versions) to clarify which fields are required and which are optional. Specifically, the 'Record Name' now indicates that it is a required field, and 'Record Tags' shows it as optional. This change improves overall user experience by providing clear instructions to the user.

Also updated the RecordStartView.axaml to shift the tags into an expandable format rather than off to the side. This enhances the user experience by offering a clearer, more organized display of tags.

Asana: https://app.asana.com/0/1203851531040615/1205310916333555/f