fix (uav): add conditional check to FollowUavCommand in FlightUavViewModel
In the FlightUavViewModel, our FollowUavCommand has been expanded with a conditional check. This allows us to ensure that if an ItemToFollow is currently selected on the Map, it will be deselected and the IsFollowed flag will be set to false.

This change improves the way our app handles user interactions with FollowUavCommand and it makes our UAV flight tracking logic more robust and straightforward. By only setting IsFollowed and ItemToFollow when a UAV vehicle is found, we avoid potential issues where the map might attempt to follow a non-existent or null item.

Asana: https://app.asana.com/0/1203851531040615/1205371221383408/f