fix: update mission status view with total and mission distances
This commit addresses changes in the mission status view regarding total & mission distances. Previously, the mission status view only displayed the total distance of all waypoints. Now, however, it is indeed enhanced to show the distance of the current mission, which is more relevant data to the user. Also, total distance now accounts for the distance from and to the home point.

This change aims to provide a clearer overview of distance-related metrics, thus assisting the user in staying aware of the drone's status and progress throughout the mission.

These modifications are reflected not only on the display elements and tooltips but also on the underlying ViewModel, where the calculation of the total distance has been adjusted to consider the home point distance. Besides, it includes changes in the localization files.

Asana: https://app.asana.com/0/1203851531040615/1205412828145386/f