feat(core): update HierarchicalStore View and ViewModel
Modified HierarchicalStore View layout to improve readability and introduced creation time display for files/folders in HierarchicalStore ViewModel. In the view, the adjustment prevents the last child of the Dockpanel from filling the whole available space for better UI. Description text is now docked to the right for clarity. Additionally, the ViewModel now includes the creation time display for files/folders when type details are fetched for FileSystemHierarchicalStoreEntry items. This provides more detailed information for the user.

Asana: https://app.asana.com/0/1203851531040615/1205628819726102/f