feat: add sorting functionality to SdrPayloadBrowserView
Added a sorting functionality to the SdrPayloadBrowserView which allows users to sort payloads by either name or date. A dropdown button has been included in the GUI to enable the users to toggle between the sorting preferences. This update is essential to enhance usability and improve the user experience, especially when dealing with a large collection of payloads. The change has been reflected in both English and Russian localizations.

Asana: https://app.asana.com/0/1203851531040615/1203719678217452/f