feat: add anchor-menu support to map page
This commit introduces a major change to the map page. A URI string is defined and a new constructor is added to the MapPageViewModel to take in a collection of IHeaderMenuItem. A transformation is added to the 'Markers' collection. This allows each IHeaderMenuItem to have its command linked to a map location, essentially adding support for an 'anchor' menu. An instance of 'HeaderAnchorsMenu' is sourced from the collection. If found, its items are set to the transformed collection.

A similar modification is made to PlaningPageViewModel, FlightPageViewModel - an exportedMenuItems parameter is added and passed on to super constructor. This ensures anchor menu support is uniform across these classes.

A new class HeaderAnchorsMenu is introduced. It exports itself as IHeaderMenuItem and has an observable collection of IHeaderMenuItem, which gets its visibility set based on if Items is null or not.

The commit also loosens the encapsulation around IHeaderMenuItem Items property to allow its mutation. This allows for the addition/removal of menu-items dynamically during program's execution. This was necessary to enable the dynamic anchor-menu functionality.

Asana: https://app.asana.com/0/1203851531040615/1203719678217443/f