fix: refactored MaterialIconConverter for simpler use
This commit simplifies the MaterialIconConverter to use 'MaterialIcon' instead of the previously used 'DrawingImage'. This results in a cleaner, more direct conversion of 'MaterialIconKind' to 'MaterialIcon', increasing readability and maintainability of the code. Removal of 'throw new NotImplementedException();' in 'ConvertBack' method results in safe backward conversion.

Asana: https://app.asana.com/0/1203851531040615/1203719678217443/f