feat (core): update ParamItemView and adding write functionality
This commit applies a redesign to the ParamItemView by removing the save param button and replacing it with a panel that is visible when the IsSynced binding returns false. This enhances the UI by simplifying the overall appearance. A write function is also added to ParamItemViewModel. This allows for data to be written to the parameter. In the ParamPageViewModel, added a function to handle scenarios where there are unsaved changes, providing options to save, don't save or close. All these changes improve synchronization and user-data interaction.

Asana: https://app.asana.com/0/1203851531040615/1205622126571075/f