feat: added unit option for drone bearing measurement
Added Degrees and Minutes unit option for drone bearing measurement and associated GUI controls.
This change allows users to view and select either Degree or Degrees and Minutes for drone bearing. The GUI changes include the addition of a ComboBox for the selection of bearing unit and update in localization strings to reflect this new option. Also, updated AsvCommonVersion to 1.12.7.

Asana: https://app.asana.com/0/1203851531040615/1205412828145367/f