fix: update sorting logic in SdrPayloadBrowserViewModel
Changes have been implemented in the SdrPayloadBrowserViewModel.cs to incorporate both ascending and descending sorting directions. This was spurred by the need to make sorting more dynamic, based on the value of IsSortByName. Additionally, an update has been introduced to trigger a download of records whenever a change is made to the 'IsSortByName' attribute.

Asana: https://app.asana.com/0/1203851531040615/1206007847454212/f