fix: remove redundant mission item addition
In the PlaningMissionRoiPointViewModel, redundant code for adding the ROI mission item to the drone mission was removed. This code was unnecessarily repeating the action of adding an ROI mission item, which is now handled by the AddRoiMissionItem method. This reduces redundancy and increases code readability.

Asana: https://app.asana.com/0/1203851531040615/1205858132447952/f