fix (core): Refactor readability of MapPageViewModel and MapRulerActionViewModel
MapPageViewModel and MapRulerActionViewModel were adjusted for better readability. The logic remained the same. Unnecessary line breaks were removed and meaningful ones were added where needed, with the aim of improving the clarity and maintainability of the code. A protected modifier was also removed from the ReadOnlyObservableCollection _markers, as it was unnecessary.

Asana: https://app.asana.com/0/1203851531040615/1204967469316979/f