feat(plugins): update plugins to include API version compatibility check
Updated the plugin system to support API version compatibility. All plugins are now required to specify their API version, which is then checked for compatibility with the application's API version. If the versions do not match, the plugin will not be loaded. The API version has also been added to the application version display for transparency.